Move the new decorators stories to the right place
